Fix nits from 232019 https://bugs.webkit.org/show_bug.cgi?id=232180
Reviewed by Saam Barati.

We only need one write barrier, since we only need to guarantee that we read the status of the write barrier before we read from the structure cache.
If we are delayed in watching the watchpoint, it does not change any of the interleavings.

* dfg/DFGAbstractInterpreterInlines.h:
(JSC::DFG::AbstractInterpreter<AbstractStateType>::executeEffects):
* dfg/DFGConstantFoldingPhase.cpp:
(JSC::DFG::ConstantFoldingPhase::foldConstants):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@284726 268f45cc-cd09-0410-ab3c-d52691b4dbfc