Add a module map file for PrivateFrameworks/WebKitLegacy https://bugs.webkit.org/show_bug.cgi?id=230735
Patch by Ian Anderson <iana@apple.com> on 2021-10-22
Reviewed by David Kilzer.

Source/WebCore:

Change a few quoted includes to framework style angle includes to
support modularization.
Make sure to define the TARGET_ macros before using them.

* platform/ios/WebItemProviderPasteboard.h:
* platform/ios/wak/WAKAppKitStubs.h:
* platform/ios/wak/WAKResponder.h:
* platform/ios/wak/WAKView.h:
* platform/ios/wak/WAKWindow.h:
* platform/ios/wak/WKContentObservation.h:
* platform/ios/wak/WebCoreThreadMessage.h:

Source/WebKitLegacy:

Add module map files for WebKitLegacy. The public module is empty
because there are no public headers. The private module map sets
PrivateHeaders as the umbrella directory because there isn't an umbrella
header already. Then it makes an explicit submodule for each header so
as to mimic the non-modular environment.

* Modules/WebKitLegacy.modulemap: Added.
* Modules/WebKitLegacy.private.modulemap: Added.
* WebKitLegacy.xcodeproj/project.pbxproj:

Source/WebKitLegacy/ios:

WebGeolocationCoreLocationProvider.h is an Objective-C++ header, but the
WebKitLegacy module should be usable by plain Objective-C clients. Add
C++ guards so that it can be included in the module.
Add a missing include to WebFixedPositionContent.h.

* Misc/WebGeolocationCoreLocationProvider.h:
* WebCoreSupport/WebFixedPositionContent.h:

Source/WebKitLegacy/mac:

Define modules for the Cocoa Touch environment. macOS can't define
modules in WebKitLegacy because it's a nested framework on that
platform.
Set up the modules verifier tool for Apple Internal (it's not supported
in the public Xcode). Don't enable it yet because some of the Safari
builders use too old of an Xcode.
Add the Apple Internal guards for NSURLDownload from WebDownload.h
NSURLDownloadSPI.h so it can be included in the module.
WebCreateFragmentInternal.h is an Objective-C++ header, but the
WebKitLegacy module should be usable by plain Objective-C clients.
C++ guards so that it can be included in the module.
Switch a few quoted includes to framework style angle includes, which
are required for modules.
Add some missing includes.

* Configurations/WebKitLegacy.xcconfig:
* DOM/DOMEventListener.h:
* DOM/DOMEventTarget.h:
* DOM/DOMNodeFilter.h:
* DOM/DOMXPathNSResolver.h:
* DOM/WebDOMOperationsPrivate.h:
* History/WebHistoryItemPrivate.h:
* Misc/NSURLDownloadSPI.h:
* Misc/WebCache.h:
* Misc/WebDownload.h:
* Misc/WebKitErrorsPrivate.h:
* Misc/WebLocalizableStrings.h:
* Misc/WebUserContentURLPattern.h:
* Plugins/Hosted/WebKitPluginHostTypes.h:
* Plugins/WebPlugin.h:
* Plugins/WebPluginContainer.h:
* Storage/WebDatabaseManagerPrivate.h:
* Storage/WebDatabaseQuotaManager.h:
* Storage/WebStorageManagerPrivate.h:
* WebCoreSupport/WebCreateFragmentInternal.h:
* WebCoreSupport/WebSecurityOriginPrivate.h:
* WebView/WebDeviceOrientation.h:
* WebView/WebDeviceOrientationProvider.h:
* WebView/WebDeviceOrientationProviderMock.h:
* WebView/WebEditingDelegatePrivate.h:
* WebView/WebFormDelegatePrivate.h:
* WebView/WebGeolocationPosition.h:
* WebView/WebResourceLoadDelegatePrivate.h:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@284736 268f45cc-cd09-0410-ab3c-d52691b4dbfc