Fix typo in MediaPlayerPrivateAVFoundationObjC::updateVideoTracks() https://bugs.webkit.org/show_bug.cgi?id=232194
Reviewed by Darin Adler.

Reset video tracks' properties when video tracks are changing rather
than the audio ones.

There should be no observable changes, resetting the audio track properties
would have been a no-op as AudioTrackPrivateAVFObjC is just a wrapper around
AVPlayerItemTrack and properties are kept in sync thorough playback.
Not calling resetPropertiesFromTrack on the video track wouldn't have mattered
as the operation is called when the track is constructed above in the
determineChangedTracksFromNewTracksAndOldItems method.
So this code only improves readability, and could be removed entirely.

* platform/graphics/avfoundation/objc/MediaPlayerPrivateAVFoundationObjC.mm:
(WebCore::MediaPlayerPrivateAVFoundationObjC::updateVideoTracks):
* platform/graphics/avfoundation/objc/VideoTrackPrivateAVFObjC.h:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@284770 268f45cc-cd09-0410-ab3c-d52691b4dbfc