ASSERT(parent->element()) triggered in Styleable::fromRenderer https://bugs.webkit.org/show_bug.cgi?id=232185
Patch by Gabriel Nava Marino <gnavamarino@apple.com> on 2021-10-26
Reviewed by Tim Nguyen and Antti Koivisto.

Source/WebCore:

The marker renderer can be set as a child of RenderMultiColumnFlowThread
instead of RenderListItem in some instances. RenderMultiColumnFlowThread is
an anonymous box and doesn't have an associated element, so we instead should
loop through the parents until we find the RenderListItem which does have an
associated element.

Test: fast/animation/css-animation-marker-crash.html

* style/Styleable.cpp:
(WebCore::Styleable::fromRenderer):

LayoutTests:

* fast/animation/css-animation-marker-crash-expected.txt: Added.
* fast/animation/css-animation-marker-crash.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@284871 268f45cc-cd09-0410-ab3c-d52691b4dbfc