REGRESSION (r282451): [iOS] Cannot override background of search inputs with 'appearance: textfield' https://bugs.webkit.org/show_bug.cgi?id=231503 rdar://84110684
Reviewed by Wenson Hsieh.

Source/WebCore:

r282451 made search inputs with 'appearance: textfield' always have the
standard text input background color. This change was necessary to fix
search input styling on sites that did not customize search input
backgrounds, but relied on 'appearance: textfield' to obtain the
appropriate styling.

However, the change also made it so that any customization to the
background color on top of the UA default was not honored. To fix,
check if the background color differs from the UA default prior to
making the adjustment.

Test: fast/forms/ios/form-control-refresh/search/textfield-appearance-background-styled.html

* rendering/RenderThemeIOS.mm:
(WebCore::RenderThemeIOS::adjustTextFieldStyle const):

LayoutTests:

* fast/forms/ios/form-control-refresh/search/textfield-appearance-background-styled-expected.txt: Added.
* fast/forms/ios/form-control-refresh/search/textfield-appearance-background-styled.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@284984 268f45cc-cd09-0410-ab3c-d52691b4dbfc