[watchOS] Fix logging https://bugs.webkit.org/show_bug.cgi?id=232423
Reviewed by Eric Carlson.

OSStatus is typedef'ed to an SInt32. On watchOS, SInt32 is typedef'ed to be a signed long.
A signed long on watchOS is 4 bytes long. On non-watchOS, SInt32 is typedef'ed to be a signed int.
So, if we want to use printf formatting strings, we have to cast the OSStatus to an int.

No new tests because there is no behavior change.

* platform/cocoa/MediaUtilities.cpp:
(WebCore::createAudioFormatDescription):
(WebCore::createAudioSampleBuffer):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@284988 268f45cc-cd09-0410-ab3c-d52691b4dbfc