DOM GPUP: Yahoo.com has fragments of wrongly-scaled content throughout https://bugs.webkit.org/show_bug.cgi?id=232356 <rdar://83884813>
Reviewed by Simon Fraser.

Source/WebCore:

* platform/graphics/ConcreteImageBuffer.h:
* platform/graphics/ImageBufferBackend.cpp:
(WebCore::ImageBufferBackend::calculateBaseTransform):
Delegate baseTransform computation to the backend, so that it
can also compute it itself when applyBaseTransformToContext() is called.

Multiply resolutionScale into the base transform.

(WebCore::ImageBufferBackend::applyBaseTransformToContext const):
Apply the base transform to the context via recordable GraphicsContext
methods, so that getCTM() returns an accurate CTM that matches the one
on the replay side.

Also, apply the device scale factor like setupContext() used to do,
but *before* setting the CTM so that we don't double-scale.
In the future we should consider making applyDeviceScaleFactor() *not*
actually scale the context, instead only touching the CG base CTM;
I went the opposite direction in r284902 in order to reduce change
to non-DisplayList clients, but this is a bit messy.

* platform/graphics/ImageBufferBackend.h:
(WebCore::ImageBufferBackend::originAtBottomLeftCorner const):
Add a virtual originAtBottomLeftCorner getter; oddly we need it both
dynamically (from ImageBufferBackend::calculateBaseTransform) and
statically (from ConcreteImageBuffer::baseTransform(), which has to
be callable before the backend is actually realized), so we have both
this virtual method and the existing static bit.

* platform/graphics/cg/ImageBufferCGBackend.cpp:
(WebCore::ImageBufferCGBackend::setupContext const): Deleted.
* platform/graphics/cg/ImageBufferCGBackend.h:
* platform/graphics/cg/ImageBufferCGBitmapBackend.cpp:
(WebCore::ImageBufferCGBitmapBackend::ImageBufferCGBitmapBackend):
* platform/graphics/cg/ImageBufferIOSurfaceBackend.cpp:
(WebCore::ImageBufferIOSurfaceBackend::ImageBufferIOSurfaceBackend):
(WebCore::ImageBufferIOSurfaceBackend::context const):
Replace setupContext() with a call to applyBaseTransformToContext(),
which supercedes it.

* platform/graphics/displaylists/DisplayListRecorder.cpp:
(WebCore::DisplayList::Recorder::Recorder):
Map the initial clip back out to base coordinates through the initialCTM;
I missed this in r284888 because the initialCTM was just a flip, but
now that it includes a scale this is critical.

Source/WebKit:

* WebProcess/GPU/graphics/cocoa/ImageBufferShareableIOSurfaceBackend.cpp:
(WebKit::ImageBufferShareableIOSurfaceBackend::backendSize const):
Implement backendSize() so that ImageBufferShareableIOSurfaceBackend
doesn't always report itself as being 0x0.
* WebProcess/GPU/graphics/cocoa/ImageBufferShareableIOSurfaceBackend.h:

LayoutTests:

* fast/transforms/interleaved-2d-transforms-with-gpu-process-expected.html: Added.
* fast/transforms/interleaved-2d-transforms-with-gpu-process.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@284991 268f45cc-cd09-0410-ab3c-d52691b4dbfc