The space glyph in the Filicudi font is not substituted by shaping in the complex text codepath, thereby breaking the visual effect of the font https://bugs.webkit.org/show_bug.cgi?id=232293
Reviewed by Darin Adler.

.:

Big thanks to Ulrike Rausch and Frank Grießhammer for contributing the test font!!!

* metadata/contributors.json:

Source/WebCore:

Filicudi is a color font where the background of the glyphs is filled with a
repeating pattern of colored vertical stripes. Each glyph has multiple
alternates which change the phase of this repeating background pattern. The
font also has shaping rules to select the right alternate glyph based on the
previous glyph, so that the phases line up correctly and you get a nice
repeating pattern throughout the whole line of text.

You can see this in action here:
https://fonts.adobe.com/fonts/tipoteca-series#fonts-section

In the complex text codepath, though, after shaping, we were clobbering the
glyph used by the space character. We were overwriting it to be the glyph
that would have been used if there were no shaping. This means that the phase
of the repeating background for the space glyph was wrong, which broke the
visual effect of the repeating background on a full line of text.

The solution is to just stop clobbering the space glyph - but continue to
clobber the space width. If we stop clobbering the space glyph, then the
repeating background starts working again. We have to continue to clobber the
spaceWidth because different parts of webkit measure text in differently-sized
chunks. For example, computing the preferred widths measures word-by-word, but
painting - which requires measuring - measures the entire run. In order to have
these different strategies measure consistently, we have to clobber the widths
of all the space characters to set them consistently. These space width
adjustments are small - they're usually due to slight kerning of characters
like "T" around a space.

Clobbering one but not the other seems a bit scary, but I think in practice it
will actually be an improvement, because:
1) These glyphs we're dealing with here are almost always invisible, so
       clobbering / not clobbering won't have much effect
2) Because the width adjustments are usually small, the chance for a mismatch
       causing broken rendering is also small.

Long term, though, we'll stop clobbering space widths, too, because IFC will
help us be more deliberate about the chunks we measure text in. So, I think
this patch is both A) an improvement upon what we're shipping today, and
B) a step in the right future direction.

Test: fast/text/complex-space-glyph.html

* platform/graphics/ComplexTextController.cpp:
(WebCore::ComplexTextController::adjustGlyphsAndAdvances):

LayoutTests:

Big thanks to Ulrike Rausch and Frank Grießhammer for contributing the test font!!!

* fast/text/complex-space-glyph-expected.html: Added.
* fast/text/complex-space-glyph.html: Added.
* fast/text/resources/Filicudi Dummy BSD2/FilicudiDummy-Barber.otf: Added.
* fast/text/resources/Filicudi Dummy BSD2/LICENSE.txt: Added.
* fast/text/resources/Filicudi Dummy BSD2/README.md: Added.
* fast/text/resources/Filicudi Dummy BSD2/color.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285009 268f45cc-cd09-0410-ab3c-d52691b4dbfc