Source/WebCore: Negative length returned by TextUtil::midWordBreak with surrogate pair https://bugs.webkit.org/show_bug.cgi?id=232053
Patch by Gabriel Nava Marino <gnavamarino@apple.com> on 2021-10-28
Reviewed by Alan Bujtas.

When the substring does not fit, the right side is supposed to be the start
of the surrogate pair if applicable. However, if the start of the surrogate
pair falls to the left of the startPosition, we will end up returning a
negative length. Instead, bail out at this point since we know that our
middle point (which becomes the right) falls before the start position.
We are not going to find a non-empty fitting content anymore in here.

* layout/formattingContexts/inline/text/TextUtil.cpp:
(WebCore::Layout::TextUtil::breakWord):

LayoutTests:
Add an additional edge case for surrogate pairs to existing tests.
https://bugs.webkit.org/show_bug.cgi?id=232053

Patch by Gabriel Nava Marino <gnavamarino@apple.com> on 2021-10-28
Reviewed by Alan Bujtas.

* fast/text/word-break-column-gap-display-flex-utf16-surrogates-expected.txt:
* fast/text/word-break-column-gap-display-flex-utf16-surrogates.html:
* fast/text/word-break-letter-spacing-utf16-surrogates-expected.txt:
* fast/text/word-break-letter-spacing-utf16-surrogates.html:
* fast/text/word-break-max-width-utf16-surrogates-expected.txt:
* fast/text/word-break-max-width-utf16-surrogates.html:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285016 268f45cc-cd09-0410-ab3c-d52691b4dbfc