Ensure synchronized rendering of incoming audio tracks https://bugs.webkit.org/show_bug.cgi?id=232375
Reviewed by Eric Carlson.

WebRTC incoming audio tracks are guaranteed to be synchronized based on the audio module clock.
Previously, we were mixing all audio tracks together, but we were not handling synchronization of tracks that have the same clock.
This can cause some lower quality audio rendering.
To prevent this, all WebRTC incoming audio tracks are now mixed together in a single track that is then mixed with all other non synchronized tracks.
To do so we introduce IncomingAudioMediaStreamTrackRendererUnit which will receive all incoming audio tracks.
It then sends the mix to the actual AudioMediaStreamTrackRendererUnit.
To ensure tight synchronization, we introduce AudioSampleDataSource::pullAvailableSampleChunk which reads data using the provided timestamp relative to the input offset.
We do a refactoring to make it easy to either use IncomingAudioMediaStreamTrackRendererUnit or AudioMediaStreamTrackRendererUnit by having a base class called BaseAudioMediaStreamTrackRendererUnit.

Manually tested.

* Modules/mediastream/libwebrtc/LibWebRTCPeerConnectionBackend.cpp:
* Modules/mediastream/libwebrtc/LibWebRTCRtpReceiverBackend.cpp:
* Modules/mediastream/libwebrtc/LibWebRTCRtpReceiverBackend.h:
* SourcesCocoa.txt:
* WebCore.xcodeproj/project.pbxproj:
* loader/EmptyClients.cpp:
* platform/audio/cocoa/AudioSampleDataSource.h:
* platform/audio/cocoa/AudioSampleDataSource.mm:
* platform/mediastream/AudioMediaStreamTrackRenderer.cpp:
* platform/mediastream/AudioMediaStreamTrackRenderer.h:
* platform/mediastream/AudioTrackPrivateMediaStream.cpp:
* platform/mediastream/RealtimeIncomingAudioSource.cpp:
* platform/mediastream/RealtimeIncomingAudioSource.h:
* platform/mediastream/cocoa/AudioMediaStreamTrackRendererCocoa.cpp:
* platform/mediastream/cocoa/AudioMediaStreamTrackRendererCocoa.h:
* platform/mediastream/cocoa/AudioMediaStreamTrackRendererUnit.h:
* platform/mediastream/cocoa/BaseAudioMediaStreamTrackRendererUnit.h: Added.
* platform/mediastream/cocoa/IncomingAudioMediaStreamTrackRendererUnit.cpp: Added.
* platform/mediastream/cocoa/IncomingAudioMediaStreamTrackRendererUnit.h: Added.
* platform/mediastream/libwebrtc/LibWebRTCAudioModule.cpp:
* platform/mediastream/libwebrtc/LibWebRTCAudioModule.h:
* platform/mediastream/libwebrtc/LibWebRTCProvider.cpp:
* platform/mediastream/libwebrtc/LibWebRTCProvider.h:
* platform/mediastream/mac/RealtimeIncomingAudioSourceCocoa.cpp:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285027 268f45cc-cd09-0410-ab3c-d52691b4dbfc