Improve error handling in sendWithAsyncReply() https://bugs.webkit.org/show_bug.cgi?id=232508
Reviewed by Darin Adler.

Source/WebKit:

* Platform/IPC/Connection.cpp:
(IPC::Connection::invalidate):
We stop delivering messages as soon as |m_valid| becomes false. It is
thus a good idea to cancel any pending async reply handlers as soon
as we set m_isValid to false. We were doing it correctly in connectionDidClose()
but not in invalidate(). The Connection destructor would eventually call
the async reply handlers. However, I have noticed when investigating a leak
that this may delay the calling of the reply handler for a long time if
something keeps the Connection alive (or forever in case of a leak).

* Platform/IPC/Connection.h:
(IPC::Connection::sendWithAsyncReply):
Return early if the connection is no longer valid and call the
completion handler right away (but asynchronously). Previously,
we'd register the async reply handler and then the call to
sendMessage() would fail.

* UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView _getContentsAsStringWithCompletionHandlerKeepIPCConnectionAliveForTesting:]):
* UIProcess/API/Cocoa/WKWebViewPrivate.h:
Add helper SPI for API test.

Tools:

Add API test coverage.

* TestWebKitAPI/Tests/WebKitCocoa/WKWebViewGetContents.mm:
(TEST):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285086 268f45cc-cd09-0410-ab3c-d52691b4dbfc