ArrayBuffer species watchpoint being invalidated doesn't mean it's not an ArrayBuffer constructor from the same global object https://bugs.webkit.org/show_bug.cgi?id=231322 <rdar://problem/84212078>
Reviewed by Yusuke Suzuki.

JSTests:

* stress/array-buffer-species-watchpoint-should-not-dictate-if-constructor-is-correct.js: Added.
(main.async v23):
(main):

Source/JavaScriptCore:

We were relying on the watchpoint to tell us if the Array species watchpoint
for 'constructor' is set to the right value. However, the watchpoint being
fired doesn't guarantee that it won't be the right value. For example, we
might not be able to install the watchpoint because the replacement watchpoint
may already be fired.

* runtime/JSArrayBufferPrototype.cpp:
(JSC::arrayBufferSpeciesConstructorSlow):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285123 268f45cc-cd09-0410-ab3c-d52691b4dbfc