Web Inspector: Display swatch popovers below the swatch by default, not on the left side https://bugs.webkit.org/show_bug.cgi?id=232577
Reviewed by Devin Rousso.

The new defaults are: below, above, left.

Displaying the swatch popover on the left side covers the relevant property name.
Displaying it below or above doesn't have this problem; below is preferred because
it doesn't hide the CSS selector of the relevant rule.

* UserInterface/Views/InlineSwatch.js:
(WI.InlineSwatch.prototype._presentPopover):
Introduce a method to remove code duplication.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285215 268f45cc-cd09-0410-ab3c-d52691b4dbfc