[macOS] MediaSession coordinator enabled in UIProcess, disabled in WebContent https://bugs.webkit.org/show_bug.cgi?id=232687 <rdar://84990898>
Reviewed by Eric Carlson.

Source/WebKit:

The default value of defaultMediaSessionCoordinatorEnabled() will be true if the UIProcess has
the correct entitlement. However, this value is not passed across the XPC boundary unless it has
been overridden, which means defaultMediaSessionCoordinatorEnabled() will control the setting
in the WebContent process, which is not entitled.

Have the WebContent process check its parent process's entitlement rather than its own when
determining the default value for defaultMediaSessionCoordinatorEnabled().

* Shared/WebPreferencesDefaultValues.cpp:
(WebKit::defaultMediaSessionCoordinatorEnabled):

Source/WTF:

Have the MediaSessionPlaylistEnabled preference use the same defaults as MediaSessionCoordinatorEnabled.

* Scripts/Preferences/WebPreferencesExperimental.yaml:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285241 268f45cc-cd09-0410-ab3c-d52691b4dbfc