[ iOS ] TestWebKitAPI.WebKitLegacy.AudioSessionCategoryIOS is crashing https://bugs.webkit.org/show_bug.cgi?id=232676
Patch by Alex Christensen <achristensen@webkit.org> on 2021-11-04
Reviewed by Eric Carlson.

In WebKitLegacy on iOS, WebCoreAVFPullDelegate.initWithPlayer is being called on the web thread,
but outputMediaDataWillChange is being called on the main run loop then accessing _player in a non-thread-safe manner.
To fix this, I call callOnMainThread to access _player on the web thread.  To fix the assertion that happens after this
fix, I made MediaPlayerPrivateAVFoundationObjC::outputMediaDataWillChange able to be called on the non-main run loop.

* platform/graphics/avfoundation/objc/MediaPlayerPrivateAVFoundationObjC.mm:
(WebCore::MediaPlayerPrivateAVFoundationObjC::outputMediaDataWillChange):
(-[WebCoreAVFPullDelegate outputMediaDataWillChange:]):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285305 268f45cc-cd09-0410-ab3c-d52691b4dbfc